reorganize jena-text unit tests: separate AbstractTestDatasetWithAnalyzer from TestDatasetWithKeywordAnalyzer
