+++ jena-text/src/test/java/org/apache/jena/query/text/TestDatasetWithKeywordAnalyzer.java
@@ -18,100 +18,21 @@
 
 package org.apache.jena.query.text;
 
-import java.io.Reader ;
-import java.io.StringReader ;
 import java.util.Arrays ;
 import java.util.HashSet ;
 import java.util.Set ;
 
-import org.apache.jena.assembler.Assembler ;
 import org.apache.jena.atlas.lib.StrUtils ;
-import org.apache.jena.query.Dataset ;
-import org.apache.jena.query.text.assembler.TextAssembler ;
-import org.apache.jena.rdf.model.Model ;
-import org.apache.jena.rdf.model.ModelFactory ;
-import org.apache.jena.rdf.model.Resource ;
-import org.junit.After ;
 import org.junit.Before ;
 import org.junit.Test ;
 
 /**
  * This class defines a setup configuration for a dataset that uses a keyword analyzer with a Lucene index.
  */
-public class TestDatasetWithKeywordAnalyzer extends AbstractTestDatasetWithTextIndexBase {
-    
-    private static final String SPEC_BASE = "http://example.org/spec#";
-    private static final String SPEC_ROOT_LOCAL = "lucene_text_dataset";
-    private static final String SPEC_ROOT_URI = SPEC_BASE + SPEC_ROOT_LOCAL;
-
-    private static String makeSpec(String analyzer, String parser) {
-        return StrUtils.strjoinNL(
-                    "prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> ",
-                    "prefix ja:   <http://jena.hpl.hp.com/2005/11/Assembler#> ",
-                    "prefix tdb:  <http://jena.hpl.hp.com/2008/tdb#>",
-                    "prefix text: <http://jena.apache.org/text#>",
-                    "prefix :     <" + SPEC_BASE + ">",
-                    "",
-                    "[] ja:loadClass    \"org.apache.jena.query.text.TextQuery\" .",
-                    "text:TextDataset      rdfs:subClassOf   ja:RDFDataset .",
-                    "text:TextIndexLucene  rdfs:subClassOf   text:TextIndex .",
-                    
-                    ":" + SPEC_ROOT_LOCAL,
-                    "    a              text:TextDataset ;",
-                    "    text:dataset   :dataset ;",
-                    "    text:index     :indexLucene ;",
-                    "    .",
-                    "",
-                    ":dataset",
-                    "    a               ja:RDFDataset ;",
-                    "    ja:defaultGraph :graph ;",
-                    ".",
-                    ":graph",
-                    "    a               ja:MemoryModel ;",
-                    ".",
-                    "",
-                    ":indexLucene",
-                    "    a text:TextIndexLucene ;",
-                    "    text:directory \"mem\" ;",
-                    "    text:queryParser " + parser + ";",
-                    "    text:entityMap :entMap ;",
-                    "    .",
-                    "",
-                    ":entMap",
-                    "    a text:EntityMap ;",
-                    "    text:entityField      \"uri\" ;",
-                    "    text:defaultField     \"label\" ;",
-                    "    text:map (",
-                    "         [ text:field \"label\" ; ",
-                    "           text:predicate rdfs:label ;",
-                    "           text:analyzer [ a " + analyzer + " ]",
-                    "         ]",
-                    "         [ text:field \"comment\" ; text:predicate rdfs:comment ]",
-                    "         ) ."
-                    );
-    }      
-    
-    public void init(String analyzer, String parser) {
-        Reader reader = new StringReader(makeSpec(analyzer, parser));
-        Model specModel = ModelFactory.createDefaultModel();
-        specModel.read(reader, "", "TURTLE");
-        TextAssembler.init();            
-        Resource root = specModel.getResource(SPEC_ROOT_URI);
-        dataset = (Dataset) Assembler.general.open(root);
-    }
-    
-    public void init(String analyzer) {
-        init(analyzer, "text:QueryParser");
-    }   
-    
+public class TestDatasetWithKeywordAnalyzer extends AbstractTestDatasetWithAnalyzer {
     @Before
     public void before() {
         init("text:KeywordAnalyzer");
-    }
-    
-    @After
-    public void after() {
-        dataset.close();
     }
     
     @Test
